+++ src/elements/displaynames/displaynames-of.spec.ts
@@ -138,4 +138,36 @@ describe('intl-displaynames-of', () => {
 
     expect(el).toHaveShadowPart('value');
   });
+
+  it('adds `lang` and `dir` on the value part element', async () => {
+    await createTestPage({
+      elements: ['intl-displaynames', 'intl-displaynames-of'],
+      html: `
+        <intl-displaynames locales="ar">
+          <intl-displaynames-of>
+            <data value="ar"></data>
+          </intl-displaynames-of>
+        </intl-displaynames>
+      `,
+    });
+    const el = document.querySelector('intl-displaynames-of') as HTMLIntlDisplayNamesOfElement;
+    const span = el.shadowRoot!.querySelector('span');
+
+    expect(span).toHaveAttribute('lang', 'ar');
+    expect(span).toHaveAttribute('dir', 'rtl');
+
+    el.providerElement!.locales = 'en';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).toHaveAttribute('lang', 'en');
+    expect(span).not.toHaveAttribute('dir');
+
+    el.providerElement!.locales = '$invalid';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).not.toHaveAttribute('lang');
+    expect(span).not.toHaveAttribute('dir');
+  });
 });