+++ src/elements/listformat/listformat-formattoparts.spec.ts
@@ -76,4 +76,37 @@ describe('intl-listformat-formattoparts', () => {
     expect(el).toHaveShadowPartsCount('literal', 1);
     expect(el).toHaveShadowPartsCount('element', 2);
   });
+
+  it('adds `lang` and `dir` on the value part element', async () => {
+    await createTestPage({
+      elements: ['intl-listformat', 'intl-listformat-formattoparts'],
+      html: `
+        <intl-listformat locales="ar">
+          <intl-listformat-formattoparts>
+            <data value="foo"></data>
+            <data value="bar"></data>
+          </intl-listformat-formattoparts>
+        </intl-listformat>
+      `,
+    });
+    const el = document.querySelector('intl-listformat-formattoparts') as HTMLIntlListFormatFormatToPartsElement;
+    const span = el.shadowRoot!.querySelector('span');
+
+    expect(span).toHaveAttribute('lang', 'ar');
+    expect(span).toHaveAttribute('dir', 'rtl');
+
+    el.providerElement!.locales = 'en';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).toHaveAttribute('lang', 'en');
+    expect(span).not.toHaveAttribute('dir');
+
+    el.providerElement!.locales = '$invalid';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).not.toHaveAttribute('lang');
+    expect(span).not.toHaveAttribute('dir');
+  });
 });