+++ src/elements/relativetimeformat/relativetimeformat-format.spec.ts
@@ -97,4 +97,37 @@ describe('intl-relativetimeformat-format', () => {
     expect(el).not.toHaveShadowPart('decimal');
     expect(el).not.toHaveShadowPart('fraction');
   });
+  
+  it('adds `lang` and `dir` on the value part element', async () => {
+    await createTestPage({
+      elements: ['intl-relativetimeformat', 'intl-relativetimeformat-format'],
+      html: `
+        <intl-relativetimeformat locales="ar">
+          <intl-relativetimeformat-format>
+            <data value="10"></data>
+            <data value="year"></data>
+          </intl-relativetimeformat-format>
+        </intl-relativetimeformat>
+      `,
+    });
+    const el = document.querySelector('intl-relativetimeformat-format') as HTMLIntlRelativeTimeFormatFormatElement;
+    const span = el.shadowRoot!.querySelector('span');
+
+    expect(span).toHaveAttribute('lang', 'ar');
+    expect(span).toHaveAttribute('dir', 'rtl');
+
+    el.providerElement!.locales = 'en';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).toHaveAttribute('lang', 'en');
+    expect(span).not.toHaveAttribute('dir');
+
+    el.providerElement!.locales = '$invalid';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(span).not.toHaveAttribute('lang');
+    expect(span).not.toHaveAttribute('dir');
+  });
 });