+++ src/elements/relativetimeformat/relativetimeformat-format.ts
@@ -1,4 +1,4 @@
-import {html} from 'lit';
+import {html, nothing} from 'lit';
 
 import AbstractRelativeTimeFormatConsumer from './abstract-relativetimeformat-consumer';
 
@@ -18,7 +18,10 @@ export default class extends AbstractRelativeTimeFormatConsumer {
     }
 
     return html`
-      <span role="none" part="value">${this.#value}</span>
+      <span role="none" part="value"
+        lang=${this.currentLang ?? nothing}
+        dir=${this.currentDir ?? nothing}
+      >${this.#value}</span>
       <span aria-hidden="true" hidden>
         <slot name="rtime"></slot>
         <slot name="unit"></slot>