+++ src/elements/relativetimeformat/relativetimeformat-formattoparts.ts
@@ -1,4 +1,5 @@
-import {html} from 'lit';
+import {html, nothing} from 'lit';
+import {map} from 'lit/directives/map.js';
 
 import HTMLIntlRelativeTimeFormatConsumerElement from './abstract-relativetimeformat-consumer';
 
@@ -20,8 +21,11 @@ export default class extends HTMLIntlRelativeTimeFormatConsumerElement {
     }
 
     return html`
-      <span role="none" part="value">
-        ${this.#value.map(part =>
+      <span role="none" part="value"
+        lang=${this.currentLang ?? nothing}
+        dir=${this.currentDir ?? nothing}
+      >
+        ${map(this.#value, part =>
             html`<span part="${part.type}" role="none">${part.value}</span>`)}
       </span>
       <span aria-hidden="true" hidden>